+++ .storybook/preview.tsx
@@ -1,41 +1,24 @@
 import type { Preview } from '@storybook/react'
-// import type { IndexEntry } from '@storybook/types'
 
 const preview: Preview = {
   parameters: {
     layout: 'centered',
     options: {
-      // TODO https://mparticle-eng.atlassian.net/browse/UNI-1214
-      // storySort: (a: IndexEntry, b: IndexEntry) => {
-      // console.log('Ordering stories', { a, b })
-      // const order = ['Documentation', 'Cell Types', 'Filters', 'Primary', 'Complex']
-      // return order.indexOf(a[1].name) - order.indexOf(b[1].name)
-      // },
       storySort: {
+        method: 'alphabetical',
+        includeNames: true,
         order: [
           'About',
-          [
-            'Introduction',
-            'Changelog',
-            'FAQ',
-            'Feedback',
-            'Component Process',
-            [
-              'Introduction',
-              'Components',
-              ['Using components', 'Change process'],
-              'Candidate Components',
-              ['Introducing new ones', 'Using existing ones', 'Promoting to a component'],
-              'Design Templates',
-            ],
-          ],
+          ['Introduction', 'Changelog', 'Feedback', 'Component Process'],
           'Foundations',
-          ['Colors', 'Typography', 'Icons', 'Errors', 'Loading'],
+          ['Typography', 'Icons', 'Color Palette'],
           'Components',
           [
             'Data Display',
             'Data Entry',
+            ['QueryItem', ['Documentation']],
             'General',
+            ['Button', ['Documentation']],
             'Feedback',
             'Layout',
             'Navigation',