+++ src/components/data-display/Popover/PopoverNotProdReady.stories.tsx
@@ -1,0 +1,39 @@
+import { Popover, type IPopoverProps } from 'src/components/data-display/Popover/Popover'
+import { type Meta, type StoryObj } from '@storybook/react'
+import { Button } from 'src/components/general/Button/Button'
+
+const meta: Meta<typeof Popover> = {
+  title: 'Components/Not Prod Ready/Data Display/Popover',
+  component: Popover,
+
+  args: {},
+}
+export default meta
+
+type Story = StoryObj<typeof Popover>
+
+const PrimaryTemplate = (args: IPopoverProps) => {
+  const content = (
+    <div>
+      <p>Content</p>
+    </div>
+  )
+
+  return (
+    <>
+      <Popover content={content} title="Title" trigger={args.trigger}>
+        <Button>{args.children}</Button>
+      </Popover>
+    </>
+  )
+}
+
+export const Hover: Story = {
+  args: { trigger: 'hover', children: 'Hover Me' },
+  render: PrimaryTemplate,
+}
+
+export const Focus: Story = {
+  args: { trigger: 'focus', children: 'Focus Me' },
+  render: PrimaryTemplate,
+}