+++ src/components/data-display/Table/Table.tsx
@@ -1,3 +1,4 @@
+import './Table.css'
 import { Table as AntTable } from 'antd'
 import { type TableProps as AntTableProps } from 'antd'
 import { type AnyObject } from 'antd/es/_util/type'
@@ -9,10 +10,13 @@ export interface ITableProps<RecordType extends AnyObject = AnyObject> extends A
 
 export type { ColumnType, ExpandableConfig, ColumnsType, TableProps }
 
-export const Table = <RecordType extends AnyObject = AnyObject>(props: ITableProps<RecordType>) => {
+export const Table = <RecordType extends AnyObject = AnyObject>({ className, ...props }: ITableProps<RecordType>) => {
   return (
     <ConfigProvider>
-      <AntTable<RecordType> {...props} />
+      <AntTable<RecordType>
+        className={`${props.sticky ? 'antd-table-override--disableStickyScroll' : ''} ${className ?? ''}`}
+        {...props}
+      />
     </ConfigProvider>
   )
 }