+++ src/components/general/Button/Button.stories.tsx
@@ -1,7 +1,7 @@
 import { Button } from 'src/components/general/Button/Button'
 import { type Meta, type StoryObj } from '@storybook/react'
 import { userEvent } from '@storybook/test'
-import { Icon } from 'src/components'
+import { Divider, Flex, Icon, Typography } from 'src/components'
 import React from 'react'
 
 const meta: Meta<typeof Button> = {
@@ -9,19 +9,13 @@ const meta: Meta<typeof Button> = {
   component: props => <Button {...props}>{props.children ?? 'Button Label'}</Button>,
 
   args: {
-    block: false,
     danger: false,
     disabled: false,
     ghost: false,
-    href: undefined,
-    htmlType: 'button',
-    icon: undefined,
     loading: false,
     shape: 'default',
     size: 'middle',
-    target: undefined,
     type: 'primary',
-    onClick: undefined,
   },
 
   argTypes: {
@@ -49,6 +43,10 @@ type Story = StoryObj<typeof Button>
 */
 
 export const Primary: Story = {
+  args: {
+    type: 'primary',
+    children: 'Create',
+  },
   play: async context => {
     const button = context.canvasElement.querySelector('button')
     if (button) {
@@ -59,9 +57,26 @@ export const Primary: Story = {
   },
 }
 
+export const PrimaryWithIcon: Story = {
+  args: {
+    type: 'primary',
+    icon: <Icon name="add" size="sm" />,
+    children: 'Add',
+  },
+}
+
 export const Default: Story = {
   args: {
     type: 'default',
+    children: 'Cancel',
+  },
+}
+
+export const DefaultWithIcon: Story = {
+  args: {
+    type: 'default',
+    icon: <Icon size="sm" name="transformation" />,
+    children: 'View Columns',
   },
 }
 
@@ -71,107 +86,41 @@ export const Dashed: Story = {
   },
 }
 
-export const Text: Story = {
+export const DashedWithIcon: Story = {
   args: {
-    type: 'text',
+    type: 'dashed',
+    icon: <Icon size="sm" name="add" />,
+    children: 'Connect Output',
   },
 }
 
 export const Link: Story = {
   args: {
     type: 'link',
+    children: 'Retry',
   },
 }
 
-export const WithDanger: Story = {
+export const LinkWithIcon: Story = {
   args: {
-    danger: true,
+    type: 'link',
+    icon: <Icon name="add" />,
+    children: 'Add Audience Criteria',
   },
 }
 
-export const Disabled: Story = {
-  args: {
-    disabled: true,
-  },
-}
-
-export const Ghost: Story = {
-  args: {
-    ghost: true,
-  },
-}
-
-export const Block: Story = {
-  args: {
-    block: true,
-  },
-}
-
-export const WithIcon: Story = {
-  args: {
-    icon: <Icon name="mpLogo" />,
-  },
-}
-
-export const WithIconSM: Story = {
-  args: {
-    type: 'default',
-    icon: <Icon name="mpLogo" size="sm" />,
-    variant: 'with-new-icon',
-  },
-}
-
-export const WithIconDefaultColorSM: Story = {
-  args: {
-    type: 'default',
-    icon: <Icon name="mpLogo" size="sm" color="default" />,
-    variant: 'with-new-icon',
-  },
-}
-
-export const RoundIconButton: Story = {
-  args: {
-    icon: <Icon name="dataPlatform" size="xl" />,
-    children: ' ',
-    type: 'default',
-    shape: 'round',
-    variant: 'with-new-icon',
-  },
-}
-export const Loading: Story = {
-  args: {
-    loading: true,
-  },
-}
-
-export const Circle: Story = {
-  args: {
-    shape: 'circle',
-  },
-}
-
-export const Round: Story = {
-  args: {
-    shape: 'round',
-  },
-}
-
-export const Large: Story = {
-  args: {
-    size: 'large',
-  },
-}
-
-export const Small: Story = {
-  args: {
-    size: 'small',
-  },
-}
-
-export const PrimaryButtonWithClick: Story = {
-  args: {
-    onClick: e => {
-      alert('Button Clicked')
-    },
+export const IconOnly: Story = {
+  render: () => {
+    return (
+      <>
+        <Flex gap="small" align="center">
+          <Button icon={<Icon name="zoomOut" />} />
+          <Typography.Text>100%</Typography.Text>
+          <Button icon={<Icon name="zoomIn" />} />
+          <Divider type="vertical" />
+          <Button icon={<Icon name="fitToScreen" />} />
+        </Flex>
+      </>
+    )
   },
 }