+++ src/components/general/FloatButton/FloatButton.stories.tsx
@@ -2,7 +2,7 @@ import { type Meta, type StoryObj } from '@storybook/react'
 import { FloatButton } from 'src/components/general/FloatButton/FloatButton'
 
 const meta: Meta<typeof FloatButton> = {
-  title: 'Components/General/FloatButton',
+  title: 'Components/Not Prod Ready/General/FloatButton',
   component: FloatButton,
 
   args: {