+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.37.1](https://github.com/mParticle/aquarium/compare/v1.37.0...v1.37.1) (2025-01-22)
+
+### Bug Fixes
+
+- double scroll on Table ([#530](https://github.com/mParticle/aquarium/issues/530)) ([2ae5258](https://github.com/mParticle/aquarium/commit/2ae5258bd777e8176a12601df3d701d63f099578))
+
+# [1.37.0](https://github.com/mParticle/aquarium/compare/v1.36.1...v1.37.0) (2025-01-17)
+
+### Features
+
+- add badge indicator ([#526](https://github.com/mParticle/aquarium/issues/526)) ([56a3a47](https://github.com/mParticle/aquarium/commit/56a3a4738a44457f1a8932b33dc3a80b0778d20b))
+- replace crosshair icon with precision ([#528](https://github.com/mParticle/aquarium/issues/528)) ([06884ae](https://github.com/mParticle/aquarium/commit/06884ae29d213efd091d6e50cea5c66f3dc95439))
+
 ## [1.36.1](https://github.com/mParticle/aquarium/compare/v1.36.0...v1.36.1) (2024-12-06)
 
 ### Bug Fixes